link-field control type
The link-field control supports relative and absolute urls. and has a
select file button to link to local ressources (images, media,
documents, ...) using finder.
It has has autocomplete for internal pages and will automatically
activate the 'open in new tab' checkbox with absolute urls.
Currently the new tab feature is a bit hackish, I will change this when
we have array support.

Create a new "Shop Item" to see it in action.